# docs/architecture/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: ArnesIA daemon API
  # 0.2.0-hs09: + GET /harnesses/{id}/nodes/{nodeId}/fuente (tab Contenido del
  # inspector, paquete inspector-drawer RF-93 — lectura confinada S2).
  # 0.3.0-hs11: + GET /version · POST /self-update (paquete boton-actualizar,
  # RF-104..107 — self-update sin sudo del binario instalado).
  # 0.4.0-hs22: + GET /portafolio · POST /portafolio/escaneos · POST /portafolio/proyectos
  # · DELETE /portafolio/arneses/{clave} (Portafolio de arneses, Slice 0 — S0-D9).
  # 0.5.0-fix-repo-self-update: + PUT /self-update/repo (bugfix 2026-07-13, RF-108/109
  # — configura el repo del self-update cuando el daemon arrancó sin --repo/ARNESIA_REPO;
  # ÚNICO endpoint que acepta un path, separado del disparador POST /self-update).
  version: 0.5.0-fix-repo-self-update
  description: >-
    Superficie HTTP/SSE del daemon `arnesia serve` (:4200). Single source: de aquí se generan los
    tipos Go (oapi-codegen) y TS (openapi-typescript). El transporte NO contiene reglas de negocio
    (docs/architecture/boundaries/dominio-independiente-de-transporte.md).


    **Confinamiento (HS-06, docs/architecture/boundaries/superficie-local-confinada.md).** Toda ruta (salvo
    `/healthz`) pasa tres gates: (1) `Host` debe ser un loopback esperado (anti DNS-rebinding); (2)
    un `Origin` de navegador debe estar en el allowlist y se refleja exacto (nunca `*`); (3) si el
...
    se documentan aquí.
servers:
  - url: http://127.0.0.1:4200/api
...
security:
  - bearerAuth: []
  # En dev (daemon sin token) la seguridad efectiva es Host+Origin; el token gate se salta.
paths:
...
  /self-update:
    post:
      operationId: selfUpdate
      summary: >-
        Self-update sin sudo (RF-104, paquete boton-actualizar): compila el árbol local
        (scripts/bundle.sh --daemon-only, cwd = repo CONFIGURADO), inspecciona el binario
        producido SIN ejecutarlo (buildinfo), instala write-tmp→rename atómico EN el directorio
        del ejecutable y agenda el re-exec POST-respuesta (RF-105). SÍNCRONO: la respuesta ES el
        reporte con el veredicto REAL de cada paso; un paso fallido corta el flujo (posteriores
        «no-corrido») y el binario instalado queda INTACTO. Seguridad (RF-106, decisión #4): bajo
        withAuth; el request NO lleva parámetros — ni rutas ni flags, el cuerpo se IGNORA; un
        self-update a la vez; huella igual = «ya-al-dia», éxito sin reinstalar.
      responses:
        "200":
          description: Reporte terminal (actualizado · ya-al-dia · fallo) con los 5 pasos.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SelfUpdateReport" }
        "409": { description: Ya hay un self-update en curso (o el daemon está por reiniciar — el lock se retiene tras «actualizado»). }
        "503": { description: "No puede NI empezar: binario no escribible (RF-102) o repo no configurado (RF-103) — el cuerpo dice cuál." }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Token de capacidad minteado por el shell Tauri (`ARNESIA_AUTH_TOKEN`). También aceptado como
        header `X-Arnesia-Token`. Ausente en dev → gate se salta (Host+Origin siguen).
    sseToken:
      type: apiKey
      in: query
      name: token
      description: Variante para el EventSource (que no puede setear headers).
  schemas:
...
    SelfUpdateReport:
      type: object
      required: [resultado, pasos]
      properties:
        resultado: { type: string, enum: [actualizado, ya-al-dia, fallo] }
        huella_nueva: { type: string, description: "huella del binario recién compilado (presente en actualizado y ya-al-dia)." }
        pasos:
          type: array
          description: "SIEMPRE los 5 pasos (verificar · build · verificar binario · instalar · reiniciar) con su veredicto REAL — la UI pinta la checklist de aquí, jamás inventa (RF-104)."
          items:
            type: object
            required: [paso, estado]
            properties:
              paso: { type: string, enum: [verificar, build, verificar binario, instalar, reiniciar] }
              estado: { type: string, enum: [ok, fallo, no-corrido, agendado], description: "«agendado» = solo el paso reiniciar tras «actualizado»: la respuesta sale ANTES del re-exec (RF-105)." }
              detalle: { type: string, description: "detalle humano; en un fallo lleva la cola del stderr." }